+++ .gitignore
@@ -1,4 +1,2 @@
 *.pyc
 *.tar
--checkpoint.ipynb
-*-checkpoint.ipynb
