--- pytorch_version/5class.names
@@ -1,5 +1,0 @@
-bird
-car
-cat
-dog
-person
