--- pytorch_version/examples/augmentation/person.names
@@ -1,1 +1,0 @@
-person
