--- pytorch_version/person.names
@@ -1,1 +1,0 @@
-person
